refactor: fix type conversion in penalty comeback logic within PenalizeValidatorsTIPSigning function

+++ eth/viction/penalty.go
@@ -133,9 +133,9 @@ func PenalizeValidatorsTIPSigning(c *posv.Posv, config *params.ChainConfig, posv
 
 	// If penalized validators has BlockSign recently, remove them from penalties
 	mapBlockHash := map[common.Hash]bool{}
-	for i := vicConfig.PenaltyComebackBlockCount - 1; i >= 0; i-- {
+	for i := int(vicConfig.PenaltyComebackBlockCount) - 1; i >= 0; i-- {
 		if len(comebacks) > 0 {
-			blockNumber := header.Number.Uint64() - i - 1
+			blockNumber := header.Number.Uint64() - uint64(i) - 1
 			header := chain.GetHeaderByNumber(blockNumber)
 			blockHash := epochBlockHashes[i]
 			if blockNumber%vicConfig.ValidatorSignInterval == 0 {
